+++ app/support/app-async-context.ts
@@ -12,6 +12,22 @@ const appAsyncContext = new AsyncLocalStorage<AppAsyncContext>();
 export const asyncContextMiddleware: Middleware = (ctx, next) =>
   appAsyncContext.run({ config: ctx.config }, next);
 
+let explicitConfig: Config | null = null;
+
+/**
+ * Allows to set the explicit config for the test purposes. It returns the
+ * rollback function that restores the previous config. The currentConfig()
+ * function will return the passed configuration until the rollback function is
+ * called.
+ */
+export function setExplicitConfig(cfg: Config): () => void {
+  const prevCfg = explicitConfig;
+  explicitConfig = cfg;
+  return () => {
+    explicitConfig = prevCfg;
+  };
+}
+
 /**
  * Returns the current application configuration for use in functions called
  * directly or indirectly by controllers. It allows to not explicitly pass
@@ -21,5 +37,5 @@ export const asyncContextMiddleware: Middleware = (ctx, next) =>
  * @returns Config
  */
 export function currentConfig(): Config {
-  return appAsyncContext.getStore()?.config ?? defaultConfig;
+  return explicitConfig ?? appAsyncContext.getStore()?.config ?? defaultConfig;
 }